+++ CHANGELOG.md
@@ -13,10 +13,13 @@ v999 (February XX, 2021)
 * Add a button to system selected component page to remove a component from the system.
 * Moved "Add a component" to a system drop down to top of selected component page.
 
+**Developer changes**
+
+* Move inclusion of `edit-component-modal.html` from `base.html` to `components/element_detail_tabs.html`.
+
 **Data changes**
 
 * Alter Element description field to be blank and none.
-
 
 v0.9.1.52 (February 16, 2021)
 -----------------------------