+++ CHANGELOG.md
@@ -4,6 +4,11 @@ GovReady-Q Release Notes
 v999 (February XX, 2021)
 ------------------------
 
+**Developer changes**
+
+* Created a ElementEditForm Django form in conjunction with some functional changes to avoid name collisions issues with component library.
+
+
 **Feature changes**
 
 * Remove a component and its statements from a system.